Prototype apimanager backup/restore - PVC-based subset

+++ pkg/reconcilers/base_reconciler.go
@@ -3,12 +3,14 @@ package reconcilers
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/3scale/3scale-operator/pkg/common"
 
 	"github.com/go-logr/logr"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -101,7 +103,11 @@ func (b *BaseReconciler) ReconcileResource(obj, desired common.KubernetesObject,
 
 	// item found successfully
 	if common.IsObjectTaggedToDelete(desired) {
-		return b.DeleteResource(desired)
+		deletePropagationPolicy := common.GetDeletePropagationPolicyAnnotation(desired)
+		if deletePropagationPolicy == nil {
+			return b.DeleteResource(desired)
+		}
+		return b.DeleteResource(desired, client.PropagationPolicy(*deletePropagationPolicy))
 	}
 
 	update, err := mutateFn(obj, desired)
@@ -116,22 +122,27 @@ func (b *BaseReconciler) ReconcileResource(obj, desired common.KubernetesObject,
 	return nil
 }
 
+func (b *BaseReconciler) GetResource(objKey types.NamespacedName, obj common.KubernetesObject) error {
+	b.Logger().Info(fmt.Sprintf("Get object '%s/%s'", strings.Replace(fmt.Sprintf("%T", obj), "*", "", 1), objKey.Name))
+	return b.Client().Get(context.TODO(), objKey, obj)
+}
+
 func (b *BaseReconciler) CreateResource(obj common.KubernetesObject) error {
-	b.Logger().Info(fmt.Sprintf("Created object %s", common.ObjectInfo(obj)))
+	b.Logger().Info(fmt.Sprintf("Created object '%s/%s'", strings.Replace(fmt.Sprintf("%T", obj), "*", "", 1), obj.GetName()))
 	return b.Client().Create(b.ctx, obj)
 }
 
 func (b *BaseReconciler) UpdateResource(obj common.KubernetesObject) error {
-	b.Logger().Info(fmt.Sprintf("Updated object %s", common.ObjectInfo(obj)))
+	b.Logger().Info(fmt.Sprintf("Updated object '%s/%s'", strings.Replace(fmt.Sprintf("%T", obj), "*", "", 1), obj.GetName()))
 	return b.Client().Update(b.ctx, obj)
 }
 
-func (b *BaseReconciler) DeleteResource(obj common.KubernetesObject) error {
-	b.Logger().Info(fmt.Sprintf("Delete object %s", common.ObjectInfo(obj)))
+func (b *BaseReconciler) DeleteResource(obj common.KubernetesObject, options ...client.DeleteOption) error {
+	b.Logger().Info(fmt.Sprintf("Delete object '%s/%s'", strings.Replace(fmt.Sprintf("%T", obj), "*", "", 1), obj.GetName()))
 	return b.Client().Delete(context.TODO(), obj)
 }
 
 func (b *BaseReconciler) UpdateResourceStatus(obj common.KubernetesObject) error {
-	b.Logger().Info(fmt.Sprintf("Updated status of object %s", common.ObjectInfo(obj)))
+	b.Logger().Info(fmt.Sprintf("Updated status of object object '%s/%s'", strings.Replace(fmt.Sprintf("%T", obj), "*", "", 1), obj.GetName()))
 	return b.Client().Status().Update(context.TODO(), obj)
 }
